added little ui

+++ app/src/main/java/com/k4dnikov/forecast/data/api/model/X.kt
@@ -2,7 +2,7 @@ package com.k4dnikov.forecast.data.api.model
 
 data class X(
     val clouds: Clouds?,
-    val dt: Int?,
+    val dt: Long?,
     val dt_txt: String?,
     val main: Main?,
     val rain: Rain?,
